Исправления после комментариев
