+++ pages.py
@@ -25,14 +25,35 @@ class BasePage:
 
 
 class LoginPage(BasePage):
-    def __init__(self, driver):
-        super().__init__(driver, timeout=60)
+    """
+    Класс для работы со страницей логина.
 
-        self.login = (By.ID, 'user-name')           #Локатор по ID для элемента строки ввода login
-        self.password = (By.ID, 'password')         #Локатор по ID для элемента строки ввода password
-        self.login_btn = (By.NAME, 'login-button')  #Локатор по Name для элемента кнопка Login
-        self.error_message = (By.CSS_SELECTOR, '[data-test="error"]')    #Локатор сообщения
-        self.page_url = 'https://www.saucedemo.com/'  # Устанавливаем URL страницы логина
+    Методы:
+        __init__(driver):
+            Инициализация драйвера и элементов страницы.
+    """
+    from selenium.webdriver.common.by import By
+    from selenium.webdriver.support.ui import WebDriverWait
+    from selenium.webdriver.support import expected_conditions as EC
+
+class LoginPage(BasePage):
+    def __init__(self, driver):
+        """
+        Локаторы для страницы логина.
+
+        Атрибуты:
+            login (tuple): Локатор строки ввода логина.
+            password (tuple): Локатор строки ввода пароля.
+            login_button (tuple): Локатор кнопки логина.
+            error_message (tuple): Локатор для отображения сообщений об ошибке.
+            page_url (str): URL страницы логина.
+        """
+        super().__init__(driver, timeout=60)
+        self.login = (By.ID, 'user-name')
+        self.password = (By.ID, 'password')
+        self.login_btn = (By.NAME, 'login-button')
+        self.error_message = (By.CSS_SELECTOR, '[data-test="error"]')
+        self.page_url = 'https://www.saucedemo.com/'
 
     def input_login(self, login: str) -> None:
         self.find_element(*self.login).send_keys(login)
